rearrange utils for clarity
