+++ packages/common/src/index.js
@@ -1,4 +1,5 @@
 export { default as Modes } from './modes'
 export { default as NuxtConfig } from './nuxt.config'
 export { default as Options } from './options'
+export * as default from './utils'
 export * from './utils'
